Add common set-version utils
