+++ packages/react-native-worklets/scripts/set-worklets-version.js
@@ -1,29 +1,17 @@
-const fs = require('fs');
 const path = require('path');
 const getVersion = require('../../../scripts/releasing').getVersion;
+const { updateVersion } = require('../../../scripts/version-utils');
 
 const packageJsonPath = path.resolve(__dirname, '../package.json');
+const jsVersionPath = path.resolve(__dirname, '../src/utils/jsVersion.ts');
 
 const { currentVersion, newVersion } = getVersion(
   process.argv,
   packageJsonPath
 );
 
-const packageJson = JSON.parse(fs.readFileSync(packageJsonPath, 'utf-8'));
-packageJson.version = newVersion;
-const newPackageJson = JSON.stringify(packageJson, null, 2) + '\n';
-fs.writeFileSync(packageJsonPath, newPackageJson, 'utf-8');
-
-const jsVersionPath = path.resolve(
-  __dirname,
-  '../src/utils/jsVersion.ts'
-);
-const before = fs.readFileSync(jsVersionPath, 'utf-8');
-const after = before.replace(
-  /jsVersion = '(.*)';/g,
-  `jsVersion = '${newVersion}';`
-);
-fs.writeFileSync(jsVersionPath, after, 'utf-8');
+// Update version using common utility
+updateVersion(packageJsonPath, jsVersionPath, newVersion);
 
 // Log the current version so it can be restored if needed.
 console.log(currentVersion);
